+++ Package@swift-5.6.swift
@@ -8,9 +8,13 @@ let package = Package(
         .library(name: "SwiftFormat", targets: ["SwiftFormat"]),
         .plugin(name: "SwiftFormatPlugin", targets: ["SwiftFormatPlugin"]),
     ],
+    dependencies: [
+        .package(url: "https://github.com/johnno1962/opaqueify.git",    .upToNextMinor(from: "1.0.19")),
+    ],
     targets: [
         .executableTarget(name: "CommandLineTool", dependencies: ["SwiftFormat"], path: "CommandLineTool"),
-        .target(name: "SwiftFormat", path: "Sources"),
+        .target(name: "SwiftFormat", dependencies: [
+            .product(name: "Opaqueifier", package: "opaqueify")], path: "Sources"),
         .testTarget(name: "SwiftFormatTests", dependencies: ["SwiftFormat"], path: "Tests"),
         .plugin(name: "SwiftFormatPlugin",
                 capability: .command(